workaround for float values

+++ cfme/tests/openstack/infrastructure/test_resources.py
@@ -16,7 +16,7 @@ pytest_generate_tests = testgen.generate([OpenstackInfraProvider],
 def test_number_of_cpu(provider, soft_assert):
     navigate_to(provider, 'Details')
     v = provider.get_detail('Properties', 'Aggregate Node CPU Resources')
-    soft_assert(int(v.split()[0]) > 0, "Aggregate Node CPU Resources is 0")
+    soft_assert(float(v.split()[0]) > 0, "Aggregate Node CPU Resources is 0")
     v = provider.get_detail('Properties', 'Aggregate Node CPUs')
     soft_assert(int(v) > 0, "Aggregate Node CPUs is 0")
     v = provider.get_detail('Properties', 'Aggregate Node CPU Cores')
@@ -26,4 +26,4 @@ def test_number_of_cpu(provider, soft_assert):
 def test_node_memory(provider):
     navigate_to(provider, 'Details')
     node_memory = provider.get_detail('Properties', 'Aggregate Node Memory')
-    assert int(node_memory.split()[0]) > 0
+    assert float(node_memory.split()[0]) > 0
